requirements.txt
Flask==3.1.2
Flask-Cors==6.0.1
google-generativeai==0.8.5
requests  # you can leave it version-free or pin to latest 2.x
python-dotenv  # same, either leave or pin to latest

# Scheduling and notifications
APScheduler
twilio